chore: type files reorganization
