+++ library/lib/types/nodes/DropNodeDate.ts
@@ -1,0 +1,8 @@
+import { DiagramNodeTypeKeys } from "@/nodes"
+
+export interface DropNodeData {
+  type: DiagramNodeTypeKeys
+  data: Record<string, unknown>
+  offsetX: number
+  offsetY: number
+}
