+++ library/lib/types/nodes/ExtraElement.ts
@@ -1,0 +1,4 @@
+export interface ExtraElement {
+  id: string
+  name: string
+}
