+++ library/lib/types/nodes/enums/ClassType.ts
@@ -1,0 +1,5 @@
+export enum ClassType {
+  Abstract = "Abstract",
+  Interface = "Interface",
+  Enumeration = "Enumeration",
+}
